more dotfiles

+++ bootstrap.sh
@@ -2,16 +2,17 @@
 
 cd "$(dirname "${BASH_SOURCE}")";
 
-git pull origin master;
+#git pull origin master;
 
 function doIt() {
 	rsync --exclude ".git/" \
 		--exclude ".DS_Store" \
 		--exclude ".osx" \
-		--exclude "bootstrap.sh" \
+                --exclude "*.sh" \
 		--exclude "README.md" \
 		--exclude "LICENSE-MIT.txt" \
 		-avh --no-perms . ~;
+	source ~/.bashrc;
 	source ~/.bash_profile;
 }
 
@@ -27,14 +28,29 @@ fi;
 unset doIt;
 
 # brew
-/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+read -p "Install brew? (y/n)" -n 1;
+echo "";
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	/usr/bin/ruby -e "$(curl -fsSL https://raw.githubusercontent.com/Homebrew/install/master/install)"
+fi;
 
 # scd
-git clone https://github.com/pavoljuhas/smart-change-directory
-sudo cp smart-change-directory/bin/scd /usr/local/bin
-scd -ar ~
+read -p "Install scd? (y/n) " -n 1;
+echo "";
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	cd ~
+	git clone https://github.com/pavoljuhas/smart-change-directory
+	sudo cp smart-change-directory/bin/scd /usr/local/bin
+	cd -
+	scd -ar ~
+fi;
 
 # Conda
-curl -O https://repo.continuum.io/archive/Anaconda3-4.4.0-MacOSX-x86_64.sh
-bash Anaconda3-4.4.0-MacOSX-x86_64.sh
+read -p "Install conda? (y/n) " -n 1;
+echo "";
+if [[ $REPLY =~ ^[Yy]$ ]]; then
+	curl -O https://repo.continuum.io/archive/Anaconda3-4.4.0-MacOSX-x86_64.sh
+	bash Anaconda3-4.4.0-MacOSX-x86_64.sh
+	rm Anaconda3-4.4.0-MacOSX-x86_64.sh
+fi;
